Bayesian Regression Models

CRRao.BayesianRegression — Type

BayesianRegression{RegressionType}

Type to represent bayesian regression models returned by fitmodel functions. This type is used internally by the package to represent all bayesian regression models. RegressionType is a Symbol representing the model class.

Linear Regression

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression, prior::Prior_Ridge, h::Float64 = 0.01, sim_size::Int64 = 10000)

Fit a Bayesian Linear Regression model on the input data with a Ridge prior.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression, prior::Prior_Laplace, h::Float64 = 0.01, sim_size::Int64 = 10000)

Fit a Bayesian Linear Regression model on the input data with a Laplace prior.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression, prior::Prior_Cauchy, sim_size::Int64 = 10000)

Fit a Bayesian Linear Regression model on the input data with a Cauchy prior.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression, prior::Prior_TDist, h::Float64 = 2.0, sim_size::Int64 = 10000)

Fit a Bayesian Linear Regression model on the input data with a t(ν) distributed prior.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression, prior::Prior_Uniform, h::Float64 = 0.01, sim_size::Int64 = 10000)

Fit a Bayesian Linear Regression model on the input data with a Uniform prior.

Logistic Regression

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::CRRaoLink, prior::Prior_Ridge, h::Float64 = 0.1, level::Float64 = 0.95, sim_size::Int64 = 10000)

Fit a Bayesian Logistic Regression model on the input data with a Ridge prior with the provided Link function.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::CRRaoLink, prior::Prior_Laplace, h::Float64 = 0.1, level::Float64 = 0.95, sim_size::Int64 = 10000)

Fit a Bayesian Logistic Regression model on the input data with a Laplace prior with the provided Link function.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::CRRaoLink, prior::Prior_Cauchy, h::Float64 = 0.1, level::Float64 = 0.95, sim_size::Int64 = 10000)

Fit a Bayesian Logistic Regression model on the input data with a Cauchy prior with the provided Link function.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::CRRaoLink, prior::Prior_TDist, h::Float64 = 1.0, level::Float64 = 0.95, sim_size::Int64 = 10000)

Fit a Bayesian Logistic Regression model on the input data with a T-Dist prior with the provided Link function.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::CRRaoLink, prior::Prior_Uniform, h::Float64 = 0.01, level::Float64 = 0.95, sim_size::Int64 = 10000)

Fit a Bayesian Logistic Regression model on the input data with a Uniform prior with the provided Link function.

Negative Binomial Regression

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::NegBinomRegression, prior::Prior_Ridge, h::Float64 = 0.1, sim_size::Int64 = 10000)

Fit a Bayesian Negative Binomial Regression model on the input data with a Ridge prior.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::NegBinomRegression, prior::Prior_Laplace, h::Float64 = 0.01, sim_size::Int64 = 10000)

Fit a Bayesian Negative Binomial Regression model on the input data with a Laplace prior.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::NegBinomRegression, prior::Prior_Cauchy, h::Float64 = 1.0, sim_size::Int64 = 10000)

Fit a Bayesian Negative Binomial Regression model on the input data with a Cauchy prior.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::NegBinomRegression, prior::Prior_TDist, h::Float64 = 1.0, sim_size::Int64 = 10000)

Fit a Bayesian Negative Binomial Regression model on the input data with a t(ν) distributed prior.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::NegBinomRegression, prior::Prior_Uniform, h::Float64 = 0.1, sim_size::Int64 = 10000)

Fit a Bayesian Negative Binomial Regression model on the input data with a Uniform prior. Ibrahim and Laud (JASA, 1990) showed that the uniform flat priors for GLMs can lead to improper posterior distributions thus making them undesirable. In such cases, the Markov Chain struggles to converge. Even if it converges, results are unreliable.

Poisson Regression

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::PoissonRegression, prior::Prior_Ridge, h::Float64 = 0.1, sim_size::Int64 = 10000)

Fit a Bayesian Poisson Regression model on the input data with a Ridge prior.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::PoissonRegression, prior::Prior_Laplace, h::Float64 = 0.1, sim_size::Int64 = 10000)

Fit a Bayesian Poisson Regression model on the input data with a Laplace prior.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression, prior::Prior_Cauchy, h::Float64 = 1.0, sim_size::Int64 = 10000)

Fit a Bayesian Poisson Regression model on the input data with a Cauchy prior.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::PoissonRegression, prior::Prior_TDist, h::Float64 = 2.0, sim_size::Int64 = 10000)

Fit a Bayesian Poisson Regression model on the input data with a t(ν) distributed prior.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::PoissonRegression, prior::Prior_Uniform, h::Float64 = 1.0, sim_size::Int64 = 10000)

Fit a Bayesian Poisson Regression model on the input data with a Uniform prior.