General Interface

Understanding the interface

CRRao exports the @fitmodel macro, which is used to train all types of models supported by the package. As of now, the macro supports the following signatures.

@fitmodel(formula, data, modelClass)
@fitmodel(formula, data, modelClass, link)
@fitmodel(formula, data, modelClass, prior)
@fitmodel(formula, data, modelClass, link, prior)

It should be noted that not all model classes support every type of signature. The parameters passed above mean the following.

The parameter formula must be a formula of type StatsModels.FormulaTerm. Any formula has an LHS and an RHS. The LHS represents the response variable, and the RHS represents the independent variables.

The parameter data must be a DataFrame. This variable represents the dataset on which the model must be trained.

modelClass represents the type of the statistical model to be used. Currently, CRRao supports four regression models, and the type of modelClass must be one of the following:

LinearRegression

LogisticRegression

NegBinomRegression

PoissonRegression

Certain model classes (like Logistic Regression) support link functions; this is represented by the link parameter. Currently four link functions are supported: Logit, Probit, Cloglog and Cauchit. So, the type of link must be one of the following:

Logit

Probit

Cloglog

Cauchit

CRRao also supports Bayesian models, and the priors to be can be specified while calling @fitmodel. Currently CRRao supports five different kinds of priors, and the type of the prior parameter must be one of the following.

Prior_Ridge

Prior_Laplace

Prior_Cauchy

Prior_TDist

Prior_Uniform

CRRao.@fitmodel — Macro

@fitmodel(formula, data, modelClass)
@fitmodel(formula, data, modelClass, link)
@fitmodel(formula, data, modelClass, prior)
@fitmodel(formula, data, modelClass, link, prior)

Macro for calling fitmodel functions to run different models.

Arguments

formula: A formula of type StatsModels.FormulaTerm.

data: A DataFrame object.

modelClass: An object of a type representing a model class.

link: The link function. Must be a subtype of CRRaoLink.

prior: An object of a type representing a prior.

Example

using CRRao, RDatasets
turnout = dataset("Zelig", "turnout")
model = @fitmodel((Vote ~ Age + Race + Income + Educate), turnout, LogisticRegression(), Logit(), Prior_Ridge())

Model Classes

CRRao.LinearRegression — Type

LinearRegression

Type representing the Linear Regression model class.

CRRao.LogisticRegression — Type

LogisticRegression

Type representing the Logistic Regression model class.

CRRao.NegBinomRegression — Type

NegBinomRegression

Type representing the Negative Binomial Regression model class.

CRRao.PoissonRegression — Type

PoissonRegression

Type representing the Poisson Regression model class.

Link functions.

CRRao.CRRaoLink — Type

CRRaoLink

Abstract type representing link functions which are used to dispatch to appropriate calls.

CRRao.Logit — Type

Logit <: CRRaoLink

A type representing the Logit link function, which is defined by the formula

\[z\mapsto \dfrac{1}{1 + \exp(-z)}\]

CRRao.Probit — Type

Probit <: CRRaoLink

A type representing the Probit link function, which is defined by the formula

\[z\mapsto \mathbb{P}[Z\le z]\]

where $Z\sim \text{Normal}(0, 1)$.

CRRao.Cloglog — Type

Cloglog <: CRRaoLink

A type representing the Cloglog link function, which is defined by the formula

\[z\mapsto 1 - \exp(-\exp(z))\]

CRRao.Cauchit — Type

Cauchit <: CRRaoLink

A type representing the Cauchit link function, which is defined by the formula

\[z\mapsto \dfrac{1}{2} + \dfrac{\text{atan}(z)}{\pi}\]

Setting Random Number Generators

CRRao.set_rng — Function

set_rng(rng)

Set the random number generator. This is useful if you want to work with reproducible results. rng must be a random number generator.

Example

using StableRNGs
CRRao.set_rng(StableRNG(1234))