CRRao

Documentation for CRRao.

CRRao.CRRao — Module

CRRao is a Julia package that implements the Statistical models. The implementation of Statistical models become straightforward for most Julia users with the help of this package. This is going to be wrapper package; leveraging the strength of wonderful Julia packages that already exists, such as StatsBase, StatsModels, Distributions,GLM, Turing, DataFrames, LinearAlgebra, etc.

CRRao is a consistent framework through which callers interact with a large suite of models. For the end-user, it reduces the cost and complexity of estimating/training statistical models. It offers convenient guidelines through which development of additional statistical models can take place in the future.

We follow framework which makes contribution to this package easy.

Note: You can read more about Prof C.R. Rao here

CRRao.BayesianRegression — Type

Type to represent bayesian regression models returned by fitmodel functions. This type is used internally by the package to represent all bayesian regression models.

CRRao.BayesianRegression — Method

BayesianRegression(::Symbol, chain)

Constructor for BayesianRegression. model can be any regression model. Used by fitmodel functions to return a frequentist regression model containers.

CRRao.CRRaoLink — Type

CRRaoLink

Abstract type representing link functions which are used to dispatch to appropriate calls.

CRRao.Cauchit — Type

Cauchit <: CRRaoLink

A type representing the Cauchit link function.

CRRao.Cloglog — Type

Cloglog <: CRRaoLink

A type representing the Cloglog link function.

CRRao.FrequentistRegression — Type

FrequentistRegression(::Symbol, model, formula, link = GLM.IdentityLink)

Constructor for FrequentistRegression. model can be any regression model. Used by fitmodel functions to return a frequentist regression model containers.

CRRao.FrequentistRegression — Type

Type to represent frequentist regression models returned by fitmodel functions. This type is used internally by the package to represent all frequentist regression models.

CRRao.Logit — Type

Logit <: CRRaoLink

A type representing the Logit link function.

CRRao.Probit — Type

Probit <: CRRaoLink

A type representing the Probit link function.

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression, prior::Prior_Uniform, h::Float64 = 0.01, sim_size::Int64 = 10000)

Fit a Bayesian Linear Regression model on the input data with a Uniform prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Linear Regression in our case.

prior: A type representing the prior. In this case, it is the Uniform prior.

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs, StatsPlots

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> df = dataset("datasets", "mtcars");

julia> container = @fitmodel(MPG ~ HP + WT + Gear, df, LinearRegression(), Prior_Uniform());

julia> plot(container.chain)

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression, prior::Prior_Cauchy, h::Float64 = 1.0, sim_size::Int64 = 10000)

Fit a Bayesian Poisson Regression model on the input data with a Cauchy prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Poisson Regression in our case.

prior: A type representing the prior. In this case, it is the Cauchy prior.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> sanction = dataset("Zelig", "sanction");

julia> container = @fitmodel(Num ~ Target + Coop + NCost, sanction, PoissonRegression(), Prior_Cauchy());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::PoissonRegression, prior::Prior_TDist, h::Float64 = 2.0, sim_size::Int64 = 10000)

Fit a Bayesian Poisson Regression model on the input data with a t(ν) distributed prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Poisson Regression in our case.

prior: A type representing the prior. In this case, it is the TDist prior.

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> sanction = dataset("Zelig", "sanction");

julia> container = @fitmodel(Num ~ Target + Coop + NCost, sanction, PoissonRegression(), Prior_TDist());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::NegBinomRegression, prior::Prior_Ridge, h::Float64 = 0.1, sim_size::Int64 = 10000)

Fit a Bayesian Negative Binomial Regression model on the input data with a Ridge prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Negative Binomial Regression in our case.

prior: A type representing the prior. In this case, it is the Ridge prior.

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> sanction = dataset("Zelig", "sanction");

julia> container = @fitmodel(Num ~ Target + Coop + NCost, sanction, NegBinomRegression(), Prior_Ridge());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::CRRaoLink, prior::Prior_Cauchy, h::Float64 = 0.1, level::Float64 = 0.95, sim_size::Int64 = 10000)

Fit a Bayesian Logistic Regression model on the input data with a Cauchy prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Logistic Regression in our case.

prior: A type representing the prior. In this case, it is the Cauchy prior.

Link: A type representing the link function to be used. Possible values are Logit(), Probit(), Cloglog() and Cauchit().

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> turnout = dataset("Zelig", "turnout");

julia> container_logit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Logit(), Prior_Cauchy());

julia> container_probit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Probit(), Prior_Cauchy());

julia> container_cloglog = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Cloglog(), Prior_Cauchy());

julia> container_cauchit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Cauchit(), Prior_Cauchy());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::PoissonRegression, prior::Prior_Uniform, h::Float64 = 1.0, sim_size::Int64 = 10000)

Fit a Bayesian Poisson Regression model on the input data with a Uniform prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Poisson Regression in our case.

prior: A type representing the prior. In this case, it is the Uniform prior.

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> sanction = dataset("Zelig", "sanction");

julia> container = @fitmodel(Num ~ Target + Coop + NCost, sanction, PoissonRegression(), Prior_Uniform());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression, prior::Prior_Ridge, h::Float64 = 0.01, sim_size::Int64 = 10000)

Fit a Bayesian Linear Regression model on the input data with a Ridge prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Linear Regression in our case.

prior: A type representing the prior. In this case, it is the Ridge prior.

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> df = dataset("datasets", "mtcars");

julia> container = @fitmodel(MPG ~ HP + WT + Gear, df, LinearRegression(), Prior_Ridge());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression, prior::Prior_Laplace, h::Float64 = 0.01, sim_size::Int64 = 10000)

Fit a Bayesian Linear Regression model on the input data with a Laplace prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Linear Regression in our case.

prior: A type representing the prior. In this case, it is the Laplace prior.

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> df = dataset("datasets", "mtcars");

julia> container = @fitmodel(MPG ~ HP + WT + Gear, df, LinearRegression(), Prior_Laplace());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression, prior::Prior_Cauchy, sim_size::Int64 = 10000)

Fit a Bayesian Linear Regression model on the input data with a Cauchy prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Linear Regression in our case.

prior: A type representing the prior. In this case, it is the Cauchy prior.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> df = dataset("datasets", "mtcars");

julia> container = @fitmodel(MPG ~ HP + WT + Gear, df, LinearRegression(), Prior_Cauchy(), 20000);

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::NegBinomRegression, prior::Prior_Laplace, h::Float64 = 0.01, sim_size::Int64 = 10000)

Fit a Bayesian Negative Binomial Regression model on the input data with a Laplace prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Negative Binomial Regression in our case.

prior: A type representing the prior. In this case, it is the Laplace prior.

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> sanction = dataset("Zelig", "sanction");

julia> container = @fitmodel(Num ~ Target + Coop + NCost, sanction, NegBinomRegression(), Prior_Laplace());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::CRRaoLink, prior::Prior_Laplace, h::Float64 = 0.1, level::Float64 = 0.95, sim_size::Int64 = 10000)

Fit a Bayesian Logistic Regression model on the input data with a Laplace prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Logistic Regression in our case.

prior: A type representing the prior. In this case, it is the Laplace prior.

Link: A type representing the link function to be used. Possible values are Logit(), Probit(), Cloglog() and Cauchit().

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> turnout = dataset("Zelig", "turnout");

julia> container_logit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Logit(), Prior_Laplace());

julia> container_probit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Probit(), Prior_Laplace());

julia> container_cloglog = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Cloglog(), Prior_Laplace());

julia> container_cauchit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Cauchit(), Prior_Laplace());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::NegBinomRegression, prior::Prior_TDist, h::Float64 = 1.0, sim_size::Int64 = 10000)

Fit a Bayesian Negative Binomial Regression model on the input data with a t(ν) distributed prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Negative Binomial Regression in our case.

prior: A type representing the prior. In this case, it is the TDist prior.

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> sanction = dataset("Zelig", "sanction");

julia> container = @fitmodel(Num ~ Target + Coop + NCost, sanction, NegBinomRegression(), Prior_TDist());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::NegBinomRegression, prior::Prior_Cauchy, h::Float64 = 1.0, sim_size::Int64 = 10000)

Fit a Bayesian Negative Binomial Regression model on the input data with a Cauchy prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Negative Binomial Regression in our case.

prior: A type representing the prior. In this case, it is the Cauchy prior.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> sanction = dataset("Zelig", "sanction");

julia> container = @fitmodel(Num ~ Target + Coop + NCost, sanction, NegBinomRegression(), Prior_Cauchy());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::NegBinomRegression, prior::Prior_Uniform, h::Float64 = 0.1, sim_size::Int64 = 10000)

Fit a Bayesian Negative Binomial Regression model on the input data with a Uniform prior. Ibrahim and Laud (JASA, 1990) showed that the uniform flat priors for GLM's can lead to improper posterior distributions thus making them undesirable. In such cases, the Markov Chain struggles to converge. Even if it converges, results are unreliable.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Negative Binomial Regression in our case.

prior: A type representing the prior. In this case, it is the Uniform prior.

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> sanction = dataset("Zelig", "sanction");

julia> container = @fitmodel(Num ~ Target + Coop + NCost, sanction, NegBinomRegression(), Prior_Uniform());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::CRRaoLink, prior::Prior_Ridge, h::Float64 = 0.1, level::Float64 = 0.95, sim_size::Int64 = 10000)

Fit a Bayesian Logistic Regression model on the input data with a Ridge prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Logistic Regression in our case.

prior: A type representing the prior. In this case, it is the Ridge prior.

Link: A type representing the link function to be used. Possible values are Logit(), Probit(), Cloglog() and Cauchit().

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> turnout = dataset("Zelig", "turnout");

julia> container_logit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Logit(), Prior_Ridge());

julia> container_probit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Probit(), Prior_Ridge());

julia> container_cloglog = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Cloglog(), Prior_Ridge());

julia> container_cauchit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Cauchit(), Prior_Ridge());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::CRRaoLink, prior::Prior_Uniform, h::Float64 = 0.01, level::Float64 = 0.95, sim_size::Int64 = 10000)

Fit a Bayesian Logistic Regression model on the input data with a Uniform prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Logistic Regression in our case.

prior: A type representing the prior. In this case, it is the Uniform prior.

Link: A type representing the link function to be used. Possible values are Logit(), Probit(), Cloglog() and Cauchit().

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> turnout = dataset("Zelig", "turnout");

julia> container_logit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Logit(), Prior_Uniform());

julia> container_probit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Probit(), Prior_Uniform());

julia> container_cloglog = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Cloglog(), Prior_Uniform());

julia> container_cauchit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Cauchit(), Prior_Uniform());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::PoissonRegression, prior::Prior_Laplace, h::Float64 = 0.1, sim_size::Int64 = 10000)

Fit a Bayesian Poisson Regression model on the input data with a Laplace prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Poisson Regression in our case.

prior: A type representing the prior. In this case, it is the Laplace prior.

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> sanction = dataset("Zelig", "sanction");

julia> container = @fitmodel(Num ~ Target + Coop + NCost, sanction, PoissonRegression(), Prior_Laplace());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression, prior::Prior_TDist, h::Float64 = 2.0, sim_size::Int64 = 10000)

Fit a Bayesian Linear Regression model on the input data with a t(ν) distributed prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Linear Regression in our case.

prior: A type representing the prior. In this case, it is the TDist prior.

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs, StatsPlots

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> df = dataset("datasets", "mtcars");

julia> container = @fitmodel(MPG ~ HP + WT + Gear, df, LinearRegression(), Prior_TDist());

julia> plot(container.chain)

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::CRRaoLink, prior::Prior_TDist, h::Float64 = 1.0, level::Float64 = 0.95, sim_size::Int64 = 10000)

Fit a Bayesian Logistic Regression model on the input data with a T-Dist prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Logistic Regression in our case.

prior: A type representing the prior. In this case, it is the T-Dist prior.

Link: A type representing the link function to be used. Possible values are Logit(), Probit(), Cloglog() and Cauchit().

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> turnout = dataset("Zelig", "turnout");

julia> container_logit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Logit(), Prior_TDist());

julia> container_probit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Probit(), Prior_TDist());

julia> container_cloglog = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Cloglog(), Prior_TDist());

julia> container_cauchit = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Cauchit(), Prior_TDist());

CRRao.fitmodel — Function

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::PoissonRegression, prior::Prior_Ridge, h::Float64 = 0.1, sim_size::Int64 = 10000)

Fit a Bayesian Poisson Regression model on the input data with a Ridge prior.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Poisson Regression in our case.

prior: A type representing the prior. In this case, it is the Ridge prior.

h: A parameter used in setting the priors.

sim_size: The number of samples to be drawn during inference.

Example

julia> using CRRao, RDatasets, StableRNGs

julia> CRRao.set_rng(StableRNG(123))
StableRNGs.LehmerRNG(state=0x000000000000000000000000000000f7)

julia> sanction = dataset("Zelig", "sanction");

julia> container = @fitmodel(Num ~ Target + Coop + NCost, sanction, PoissonRegression(), Prior_Ridge());

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression)

Fit an OLS Linear Regression model on the input data. Uses the lm method from the GLM package under the hood.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Linear Regression in our case.

julia> using CRRao, RDatasets, StatsPlots

julia> df = dataset("datasets", "mtcars");

julia> container = @fitmodel(MPG ~ HP + WT + Gear, df, LinearRegression());

julia> coeftable(container);

julia> sigma(container);

julia> loglikelihood(container);

julia> aic(container);

julia> bic(container);

julia> r2(container);

julia> adjr2(container);

julia> predict(container);

julia> residuals(container);

julia> cooksdistance(container);

julia> plot(cooksdistance(container));

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::Cauchit)

Fit a Logistic Regression model on the input data using the Cauchit link. Uses the glm method from the GLM package under the hood. ```

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::Cloglog)

Fit a Logistic Regression model on the input data using the Cloglog link. Uses the glm method from the GLM package under the hood. ```

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::Logit)

Fit a Logistic Regression model on the input data using the Logit link. Uses the glm method from the GLM package under the hood.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Logistic Regression in this case.

Link: A type to specify the link function to be used, which is Logit in this case.

julia> using CRRao, RDatasets

julia> turnout = dataset("Zelig", "turnout");

julia> container = @fitmodel(Vote ~ Age + Race + Income + Educate, turnout, LogisticRegression(), Logit());

julia> coeftable(container);

julia> loglikelihood(container);

julia> aic(container);

julia> bic(container);

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::Probit)

Fit a Logistic Regression model on the input data using the Probit link. Uses the glm method from the GLM package under the hood.

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::NegBinomRegression)

Fit a Negative Binomial Regression model on the input data (with the default link function being the Log link). Uses the glm method from the GLM package under the hood.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Negative Binomial Regression in this case.

Example

julia> using CRRao, RDatasets

julia> sanction = dataset("Zelig", "sanction");

julia> container = @fitmodel(Num ~ Target + Coop + NCost, sanction, NegBinomRegression());

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::PoissonRegression)

Fit a Poisson Regression model on the input data (with the default link function being the Log link). Uses the glm method from the GLM package under the hood.

Arguments

formula: A formula term representing dependencies between the columns in the dataset.

data: The dataset.

modelClass: Object representing the type of regression, which is Poisson Regression in this case.xc

Example

julia> using CRRao, RDatasets

julia> sanction = dataset("Zelig", "sanction");

julia> container = @fitmodel(Num ~ Target + Coop + NCost, sanction, PoissonRegression());

CRRao.set_rng — Method

using StableRNGs
CRRao.set_rng(StableRNG(1234))

CRRao.@fitmodel — Macro

   @fitmodel(formula, args...)

Macro for calling fitmodel without using @formula to create the formula.

Example

using CRRao, RDatasets
sanction = dataset("Zelig", "sanction")
model = @fitmodel(Num ~ Target + Coop + NCost, sanction, NegBinomRegression())