Frequentist Regression Models

CRRao.FrequentistRegression — Type

FrequentistRegression{RegressionType}

Type to represent frequentist regression models returned by fitmodel functions. This type is used internally by the package to represent all frequentist regression models. RegressionType is a Symbol representing the model class.

Linear Regression

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LinearRegression)

Fit an OLS Linear Regression model on the input data. Uses the lm method from the GLM package under the hood. Returns an object of type FrequentistRegression{:LinearRegression}.

Logistic Regression

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::Logit)

Fit a Logistic Regression model on the input data using the Logit link. Uses the glm method from the GLM package under the hood. Returns an object of type FrequentistRegression{:LogisticRegression}.

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::Probit)

Fit a Logistic Regression model on the input data using the Probit link. Uses the glm method from the GLM package under the hood. Returns an object of type FrequentistRegression{:LogisticRegression}.

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::Cloglog)

Fit a Logistic Regression model on the input data using the Cloglog link. Uses the glm method from the GLM package under the hood. Returns an object of type FrequentistRegression{:LogisticRegression}.

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::LogisticRegression, Link::Cauchit)

Fit a Logistic Regression model on the input data using the Cauchit link. Uses the glm method from the GLM package under the hood. Returns an object of type FrequentistRegression{:LogisticRegression}.

Negative Binomial Regression

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::NegBinomRegression)

Fit a Negative Binomial Regression model on the input data (with the default link function being the Log link). Uses the glm method from the GLM package under the hood. Returns an object of type FrequentistRegression{:NegativeBinomialRegression}.

Poisson Regression

CRRao.fitmodel — Method

fitmodel(formula::FormulaTerm, data::DataFrame, modelClass::PoissonRegression)

Fit a Poisson Regression model on the input data (with the default link function being the Log link). Uses the glm method from the GLM package under the hood. Returns an object of type FrequentistRegression{:PoissonRegression}.

Extended functions from StatsAPI.jl

CRRao.coeftable — Method

coeftable(container::FrequentistRegression)

Table of coefficients and other statistics of the model. Extends the coeftable method from StatsAPI.jl.

Example

using CRRao, RDatasets

# Get the dataset
mtcars = dataset("datasets", "mtcars")

# Train the model
container = @fitmodel(MPG ~ HP + WT + Gear, mtcars, LinearRegression())

# Get table of coefficients
coeftable(container)

CRRao.r2 — Method

r2(container::FrequentistRegression)

Coeffient of determination. Extends the r2 method from StatsAPI.jl.

Example

using CRRao, RDatasets

# Get the dataset
mtcars = dataset("datasets", "mtcars")

# Train the model
container = @fitmodel(MPG ~ HP + WT + Gear, mtcars, LinearRegression())

# Get r2
r2(container)

CRRao.adjr2 — Method

adjr2(container::FrequentistRegression)

Adjusted coeffient of determination. Extends the adjr2 method from StatsAPI.jl.

Example

using CRRao, RDatasets

# Get the dataset
mtcars = dataset("datasets", "mtcars")

# Train the model
container = @fitmodel(MPG ~ HP + WT + Gear, mtcars, LinearRegression())

# Get adjr2
adjr2(container)

CRRao.loglikelihood — Method

loglikelihood(container::FrequentistRegression)

Log-likelihood of the model. Extends the loglikelihood method from StatsAPI.jl.

Example

using CRRao, RDatasets

# Get the dataset
mtcars = dataset("datasets", "mtcars")

# Train the model
container = @fitmodel(MPG ~ HP + WT + Gear, mtcars, LinearRegression())

# Get loglikelihood
adjr2(container)

CRRao.aic — Method

aic(container::FrequentistRegression)

Akaike's Information Criterion. Extends the aic method from StatsAPI.jl.

Example

using CRRao, RDatasets

# Get the dataset
mtcars = dataset("datasets", "mtcars")

# Train the model
container = @fitmodel(MPG ~ HP + WT + Gear, mtcars, LinearRegression())

# Get aic
aic(container)

CRRao.bic — Method

bic(container::FrequentistRegression)

Bayesian Information Criterion. Extends the bic method from StatsAPI.jl.

Example

using CRRao, RDatasets

# Get the dataset
mtcars = dataset("datasets", "mtcars")

# Train the model
container = @fitmodel(MPG ~ HP + WT + Gear, mtcars, LinearRegression())

# Get bic
bic(container)

Missing docstring.

Missing docstring for sigma(container::FrequentistRegression). Check Documenter's build log for details.

CRRao.predict — Method

predict(container::FrequentistRegression)

Predicted response of the model. Extends the predict method from StatsAPI.jl.

Example

using CRRao, RDatasets

# Get the dataset
mtcars = dataset("datasets", "mtcars")

# Train the model
container = @fitmodel(MPG ~ HP + WT + Gear, mtcars, LinearRegression())

# Get predicted response
predict(container)

CRRao.residuals — Method

residuals(container::FrequentistRegression)

Residuals of the model. Extends the residuals method from StatsAPI.jl.

Example

using CRRao, RDatasets

# Get the dataset
mtcars = dataset("datasets", "mtcars")

# Train the model
container = @fitmodel(MPG ~ HP + WT + Gear, mtcars, LinearRegression())

# Get residuals
residuals(container)

CRRao.cooksdistance — Method

cooksdistance(container::FrequentistRegression)

Compute Cook's distance for each observation in a linear model. Extends the cooksdistance method from StatsAPI.jl.

Example

using CRRao, RDatasets

# Get the dataset
mtcars = dataset("datasets", "mtcars")

# Train the model
container = @fitmodel(MPG ~ HP + WT + Gear, mtcars, LinearRegression())

# Get vector of Cook's distances
cooksdistance(container)